Alter code for new decipher, english very slow and not working properly yet
