--- code/main.py
@@ -1,10 +1,0 @@
-from code import decipher_api
-
-
-def main():
-    print("--- Initialize the application. ---")
-    decipher_api.init(8080)
-
-if __name__ == '__main__':
-    print("Running main")
-    main()
